# funkey-bite-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Funkey Bite API Contract",
    "version": "2026-05-10",
    "description": "Canonical route contract consumed by the admin and web frontends."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local API server"
    }
  ],
  "paths": {
...
    "/api/v1/orders/track/{orderNumber}": {"get": {"operationId": "trackOrder", "summary": "Track order by order number", "parameters": [{"name": "orderNumber", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Tracked order"}}}},
    "/api/v1/order/track/{phone}/{orderNumber}": {"get": {"operationId": "trackOrderPublic", "summary": "Track order by phone and order number", "parameters": [{"name": "phone", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "orderNumber", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Tracked order"}}}},
...
    "/api/v1/admin/orders/{id}/status": {"patch": {"operationId": "updateOrderStatus", "summary": "Update order status", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}], "responses": {"200": {"description": "Order status update acknowledgement"}}}},
...
  }
}